backend/requirements.txt
# FastAPI and Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Existing ML dependencies (already installed)
pyyaml>=6.0
pydantic>=2.5.0

# Background tasks
asyncio>=3.4.3
